requirements.txt
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
google-api-core==1.8.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
idna==2.8
pkg-resources==0.0.0
protobuf==3.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
yapf==0.26.0
